userportal,webadmin: UI redesign
- Updated tab look and feel to be PF3 menus
- Removed System tree as it has no place in a PF application.
- Moved alerts/tasks/events to top
- Moved bookmarks to search panel
- Moved tags to configure menu
- Menu on the left side.
- Fixed z-index of modals to match bootstrap values.

+++ frontend/webadmin/modules/gwt-common/src/main/java/org/ovirt/engine/ui/common/presenter/DynamicTabProxy.java
@@ -34,7 +34,7 @@ public abstract class DynamicTabProxy<T extends DynamicTabPresenter<?, ?>> exten
                 Type<ChangeTabHandler> changeTabEventType,
                 String label, float priority, String historyToken, Align align) {
             bind(placeManager, eventBus);
-            this.tabData = new DynamicTabData(label, priority, historyToken, align);
+            this.tabData = new DynamicTabData(label, (int)priority, historyToken, align);
             this.targetHistoryToken = historyToken;
             this.requestTabsEventType = requestTabsEventType;
             this.changeTabEventType = changeTabEventType;
